fix: organize hook & nav ui
